% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frogsfunction.R
\name{frogsfunction}
\alias{frogsfunction}
\title{frogsfunction}
\usage{
frogsfunction(data, predictor)
}
\arguments{
\item{data}{caracter}

\item{predictor}{caracter}
}
\value{
caracter
}
\description{
frogsfunction
}
\examples{
frogsfunction()
}
