% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlat_to_utm.R
\name{longlat_to_utm}
\alias{longlat_to_utm}
\title{longlat_to_utm}
\usage{
longlat_to_utm(data, longlat_crs, utm_crs)
}
\arguments{
\item{data}{a dataframe}

\item{longlat_crs}{a variable in the dataframe}

\item{utm_crs}{a variable in the dataframe}
}
\value{
a dataframe
}
\description{
longlat_to_utm
}
\examples{
longlat_to_utm(orfrogs, "+proj=longlat +datum=WGS84", "+proj=utm +zone=10")
}
