% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregonfrogs-package.R
\docType{package}
\name{oregonfrogs-package}
\alias{oregonfrogs}
\alias{oregonfrogs-package}
\title{oregonfrogs: Oregon Spotted A Frog (Rana Pretiosa)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Oregon Frogs Rana Pretiosa dataset contains information about radio telemetry frequency used to study the frogs' habitat. The survey run from late September to November 2018. There are 311 observations and 16 variables.
}
\keyword{internal}
