% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_file.R
\name{path_to_file}
\alias{path_to_file}
\title{Get file path to `oregonfrogs_raw.csv` file}
\source{
This function is adapted from `readxl::readxl_example()`.
}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{Name of file in quotes with extension;
`"oregonfrogs_raw.csv"` will work.
If `NULL`, the example files will be listed.}
}
\description{
oregonfrogs comes bundled with csv files in its `inst/extdata`
directory. This function make them easy to access.
}
\details{
The data in each file is the same as \code{\link{oregonfrogs_raw}},
but without the variable types.
}
\examples{
path_to_file()
path_to_file("penguins_raw.csv")
head(read.csv(path_to_file("penguins.csv")))
}
