% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_to_longlat.R
\name{utm_to_longlat}
\alias{utm_to_longlat}
\title{utm_to_longlat}
\usage{
utm_to_longlat(data, utm_crs, longlat_crs)
}
\arguments{
\item{data}{a dataframe}

\item{utm_crs}{a variable in the dataframe}

\item{longlat_crs}{a variable in the dataframe}
}
\value{
a dataframe
}
\description{
utm_to_longlat
}
\examples{
utm_to_longlat(oregonfrogs_raw, "+proj=utm +zone=10", "+proj=longlat +datum=WGS84")
}
